feat: Refactor SurveyForm & TitleBar components, enhance UI styling
- Separated SurveyForm and TitleBar from MovieSurveyForm for better readability and maintainability
- Improved UI styling to enhance visual appeal and user experience
