+++ src/components/MovieSurveyForm.jsx
@@ -1,6 +1,7 @@
-import "../App.css";
+import "../assets/App.css";
 import { useState } from "react";
-import SuccessWindow from "./Succeswindow";
+import SuccessWindow from "./FormComponents/Succeswindow";
+import SurveyForm from "./FormComponents/SurveyForm";
 
 function MovieSurveyForm() {
   const [formInfo, setFormInfo] = useState({
@@ -19,7 +20,6 @@ function MovieSurveyForm() {
 
   const [isSubmitted, setIsSubmitted] = useState(false);
 
-
   const handleChange = (event) => {
     const { name, value } = event.target;
     setFormInfo({
@@ -30,6 +30,12 @@ function MovieSurveyForm() {
 
   const handleReset = () => {
     setFormInfo({
+      name: "",
+      email: "",
+      selectedOption: "",
+      opinion: "",
+    });
+    setErrors({
       name: "",
       email: "",
       selectedOption: "",
@@ -71,7 +77,6 @@ function MovieSurveyForm() {
 
   const handleSubmit = (event) => {
     event.preventDefault();
-    event.preventDefault();
     if (validateForm()) {
       setIsSubmitted(true);
     }
@@ -94,8 +99,7 @@ function MovieSurveyForm() {
   //Movie Survey Form
   return (
     <div className="min-h-screen flex gap-y-4 justify-center">
-      <div className="bg-white w-full max-w-lg rounded-md shadow-md">
-        {/* If isSubmitted is TRUE, show success window; otherwise, show form */}
+      {/* If isSubmitted is TRUE, show success window; otherwise, show form */}
       {isSubmitted ? (
         <SuccessWindow
           formInfo={formInfo}
@@ -106,126 +110,15 @@ function MovieSurveyForm() {
           }}
         />
       ) : (
-        <>
-        {/* Title */}
-        <div class="bg-gradient-to-r from-purple-700 to-blue-600 flex flex-col space-y-1.5 p-6">
-          <div className="text-2xl font-semibold mb-6 text-white flex items-center gap-2">
-            Movie Survey
-          </div>
-        </div>
-
-        {/* Form */}
-        <form
-          noValidate
-          onSubmit={handleSubmit}
-          className="flex flex-col space-y-4 m-4 p-2"
-        >
-          {/* Name */}
-          <div className="mb-4">
-            <label
-              htmlFor="name"
-              className="block font-medium text-gray-700 mb-1 ml-0"
-            >
-              ชื่อ <span className="text-red-500">*</span>
-            </label>
-            <input
-              className="mt-1 w-full border border-gray-300 rounded p-2 focus:outline-none focus:border-purple-500"
-              type="text"
-              id="name"
-              name="name"
-              placeholder="กรุณากรอกชื่อของคุณ"
-              value={formInfo.name}
-              onChange={handleChange}
-            />
-            {errors.name && (
-              <p className="text-red-500 text-sm mt-1">{errors.name}</p>
-            )}
-          </div>
-
-          {/* Email */}
-          <div>
-            <label htmlFor="email" className="block font-medium text-gray-700">
-              อีเมล <span className="text-red-500">*</span>
-            </label>
-            <input
-              className="mt-1 w-full border border-gray-300 rounded p-2 focus:outline-none focus:border-purple-500"
-              type="email"
-              id="email"
-              name="email"
-              placeholder="example@email.com"
-              value={formInfo.email}
-              onChange={handleChange}
-            />
-            {errors.email && (
-              <p className="text-red-500 text-sm mt-1">{errors.email}</p>
-            )}
-          </div>
-
-          {/* Movie lists */}
-          <div>
-            <p className="font-medium text-gray-700 mb-2">เลือกหนังที่คุณชอบ</p>
-            <div className={`p-2 rounded-md ${errors.selectedOption ? "border border-red-500" : "border border-transparent"}`}>
-            {movies.map((movie) => (
-              <label
-                key={movie.title}
-                className="flex items-center space-x-2 mb-2 p-2 hover:bg-gray-100 hover:rounded-sm"
-              >
-                <input
-                  className="focus:ring-purple-500 "
-                  type="radio"
-                  name="selectedOption"
-                  value={movie.title}
-                  checked={formInfo.selectedOption === movie.title}
-                  onChange={handleChange}
-                />
-                <div className="ml-1">
-                <span className="block">{`${movie.title} (${movie.year}) `}</span> 
-                <span className="block text-gray-500">{`Director: ${movie.director}`}</span>
-                </div>
-              </label>
-            ))}
-            </div>
-            {errors.selectedOption && (
-              <p className="text-red-500 text-sm mt-1">
-                {errors.selectedOption}
-              </p>
-            )}
-          </div>
-
-          {/* Comment */}
-          <div>
-            <label>ความคิดเห็นเกี่ยวกับหนัง</label>
-            <textarea
-              name="opinion"
-              rows="4"
-              value={formInfo.opinion}
-              onChange={handleChange}
-              placeholder="พิมพ์ความคิดเห็นของคุณที่นี้..."
-              className="mt-1 w-full border border-gray-300 rounded p-2 focus:outline-none focus:border-purple-500"
-            ></textarea>
-          </div>
-
-          {/* Buttons */}
-          <div className="flex space-x-2">
-            <button
-              type="button"
-              onClick={handleReset}
-              className="px-4 py-2 bg-gray-200 text-gray-700 rounded hover:bg-gray-300"
-            >
-              รีเซ็ต
-            </button>
-
-            <button
-              type="submit"
-              className="px-4 py-2 bg-purple-600 text-white rounded hover:bg-purple-700"
-            >
-              ส่งแบบสำรวจ
-            </button>
-          </div>
-        </form>
-        </>
-        )}
-      </div>
+        <SurveyForm
+          formInfo={formInfo}
+          errors={errors}
+          movies={movies}
+          handleChange={handleChange}
+          handleReset={handleReset}
+          handleSubmit={handleSubmit}
+        />
+      )}
     </div>
   );
 }
